test2

+++ src/string_buffer.c
@@ -158,9 +158,9 @@ void string_buffer_release(struct StringBuffer *buffer)
 
   if (buffer->value != NULL)
   {
-    buffer->value[buffer->content_size] = 0;                                                                      //todo ???
-    printf("free, size: %u srln: %u content: %si\n", buffer->content_size, strlen(buffer->value), buffer->value); //todo remo
-    //free(buffer->value);
+    buffer->value[buffer->content_size] = 0;//todo ???
+    printf("free, size: %d srln: %d content: %si\n",(int) buffer->content_size,(int) strlen(buffer->value), buffer->value);//todo remo
+//    free(buffer->value);
     buffer->value = NULL;
   }
 
@@ -393,7 +393,7 @@ bool _set_capacity(struct StringBuffer *buffer, const size_t size)
   buffer->value[buffer->content_size] = 0; // ensure we null at end of content
 
   buffer->max_size = size;
-  printf("size: %u strln: %u content: %s\n", buffer->content_size, strlen(buffer->value), buffer->value);//todo remo
+  printf("size: %d strln: %d content: %s\n",(int) buffer->content_size,(int) strlen(buffer->value), buffer->value);//todo remo
   buffer->value = realloc(buffer->value, buffer->max_size * sizeof(char));
 
   // put null at end
